requirements.txt
fastapi
uvicorn
pillow
numpy
jinja2
python-multipart
joblib
tensorflow